CoolBlog/requirements.txt
appdirs==1.4.3
diff-match-patch==20121119
Django==1.9.8
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.5.1
django-import-export==1.0.0
django-js-asset==1.0.0
django-pure-pagination==0.3.0
django-reversion==2.0.13
django-simple-captcha==0.4.6
django-tinymce==2.7.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
jdcal==1.4
jieba==0.39
odfpy==1.3.6
olefile==0.44
openpyxl==2.5.3
packaging==16.8
Pillow==4.1.1
PyMySQL==0.7.11
pyparsing==2.2.0
pytz==2018.4
PyYAML==3.12
six==1.10.0
tablib==0.12.1
tornado==5.0.2
unicodecsv==0.14.1
Whoosh==2.7.4
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
